+++ tests/testthat/test_importPolarized.r
@@ -2,18 +2,20 @@
 test_that("error on arguments in importPolarized", {
   expect_error(
     object = importPolarized(
-      file = system.file("extdata", "data7x10.txt", package = "diemr"),
+      files = system.file("extdata", "data7x10.txt", package = "diemr"),
       changePolarity = rep(FALSE, 3),
-      ChosenInds = 1:7
+      ChosenInds = 1:7,
+      ChosenSites = "all"
     ),
-    regexp = "but changePolarity has length"
+    regexp = "does not have the same length"
   )
 
   expect_error(
     object = importPolarized(
-      file = system.file("extdata", "data7x10.txt", package = "diemr"),
+      files = system.file("extdata", "data7x10.txt", package = "diemr"),
       changePolarity = rep(FALSE, 10),
-      ChosenInds = 1:8
+      ChosenInds = 1:8,
+      ChosenSites = "all"
     ),
     regexp = "contains fewer individuals"
   )
@@ -21,44 +23,50 @@ test_that("error on arguments in importPolarized", {
 
 
 test_that("correct solution of importPolarized", {
+  
+  local_edition(3)
+  
   expect_equal(
     object = importPolarized(
-      file = system.file("extdata", "data7x3.txt", package = "diemr"),
+      files = system.file("extdata", "data7x3.txt", package = "diemr"),
       changePolarity = rep(FALSE, 3),
-      ChosenInds = 1:6
+      ChosenInds = 1:6,
+      ChosenSites = "all"
     ),
-    expected = c(
+    expected = matrix(c(
       "0", "0", "1", "1", "2", "2", "1", "2", "1",
       "0", "0", "0", "0", "2", "2", "2", "1", "_"
-    ),
+    ), ncol = 3),
     ignore_attr = TRUE
   )
 
   expect_equal(
     object = importPolarized(
-      file = system.file("extdata", "data7x3.txt", package = "diemr"),
+      files = system.file("extdata", "data7x3.txt", package = "diemr"),
       changePolarity = c(FALSE, FALSE, TRUE),
-      ChosenInds = 1:6
+      ChosenInds = 1:6,
+      ChosenSites = "all"
     ),
-    expected = c(
+    expected = matrix(c(
       "0", "0", "1", "1", "2", "2",
       "1", "2", "1", "0", "0", "0",
       "2", "0", "0", "0", "1", "_"
-    ),
+    ), ncol = 3),
     ignore_attr = TRUE
   )
 
   expect_equal(
     object = importPolarized(
-      file = system.file("extdata", "data7x3.txt", package = "diemr"),
+      files = system.file("extdata", "data7x3.txt", package = "diemr"),
       changePolarity = c(TRUE, TRUE, TRUE),
-      ChosenInds = 1:6
+      ChosenInds = 1:6,
+      ChosenSites = "all"
     ),
-    expected = c(
+    expected = matrix(c(
       "2", "2", "1", "1", "0", "0",
       "1", "0", "1", "2", "2", "2",
       "2", "0", "0", "0", "1", "_"
-    ),
+    ), ncol = 3),
     ignore_attr = TRUE
   )
 })
@@ -67,9 +75,10 @@ test_that("correct solution of importPolarized", {
 test_that("correct dimnames in importPolarized", {
   expect_equal(
     object = importPolarized(
-      file = system.file("extdata", "data7x3.txt", package = "diemr"),
+      files = system.file("extdata", "data7x3.txt", package = "diemr"),
       changePolarity = c(TRUE, TRUE, TRUE),
-      ChosenInds = c(1:2, 4:6)
+      ChosenInds = c(1:2, 4:6),
+      ChosenSites = "all"
     ),
     expected = matrix(c(
       "2", "2", "1", "0", "0",
@@ -83,9 +92,10 @@ test_that("correct dimnames in importPolarized", {
 
   expect_equal(
     object = importPolarized(
-      file = system.file("extdata", "data7x3.txt", package = "diemr"),
+      files = system.file("extdata", "data7x3.txt", package = "diemr"),
       changePolarity = c(TRUE, TRUE, TRUE),
-      ChosenInds = c(1, 4, 2, 3, 6, 5)
+      ChosenInds = c(1, 4, 2, 3, 6, 5),
+      ChosenSites = "all"
     ),
     expected = matrix(c(
       "2", "1", "2", "1", "0", "0",
@@ -96,4 +106,23 @@ test_that("correct dimnames in importPolarized", {
     dimnames = list(c("1", "4", "2", "3", "6", "5"), c("m1", "m2", "m3"))
     )
   )
+  
+    expect_equal(
+    object = importPolarized(
+      files = c(system.file("extdata", "data7x3.txt", package = "diemr"),
+                system.file("extdata", "data7x10.txt", package = "diemr")),
+      changePolarity = c(TRUE, rep(FALSE, 12)),
+      ChosenInds = c(1, 5),
+      ChosenSites = c(TRUE, FALSE, FALSE, TRUE, rep(FALSE, 9))
+    ),
+    expected = matrix(c(
+      "2", "0", 
+      "1", "2"      
+    ),
+    ncol = 2,
+    dimnames = list(c("1", "5"), c("m1", "m4"))
+    )
+  )
+
+  
 })